tweak: change config loading strategy
The configuration handler will now look for the command-line value of `SchwiConfigPath` and attempt to form a valid path from it to then load (or create) a config file. When a config path isn't specified, the handler will use ./SchwiConfig.json as the configuration file, wherever that happens to be.